Old Version
This is what Untitled Symbol Map Tool looked like before I revamped it to use an actual Code::Blocks Project.  This is a horrendous mess that should never be looked to as an example, and I'm only uploading this so I can point and laugh at my own failures of my past.